+++ subpackages/lib/source/odood/lib/python/odoo.d
@@ -46,10 +46,10 @@ string suggestPythonVersion(in OdooSerie serie) {
     if (serie == OdooSerie(17))
         return "3.11.15";
     if (serie == OdooSerie(18))
-        return "3.12.13";
+        return "3.12.14";
     if (serie == OdooSerie(19))
-        return "3.12.13";
-    return "3.12.13";
+        return "3.12.14";
+    return "3.12.14";
 }